+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## [Unreleased]
+* Fix: local stations not being returned correctly
 
 ## [2.0.5] - 2026-08-29
 * Fix: add the schedule menu back for international listeners